+++ CHANGELOG.md
@@ -6,4 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 
 ### Added
-- Initial release for the package.
+- Ability to view core commands for ST & SM with documentation.
+- Ability to view plugin/package commands for ST with ability to navigate to source.
+- Various settings to filter based on host/command type etc.
+- Key binding to auto open the documentation panel for core commands.
+- Using modifier keys to copy command signature to clipboard.